feat - 로그인, 회원가입, 학교 인증, 정보 조회 mysql 연결

+++ src/main/java/com/example/cupcycle/repository/CafeRepository.java
@@ -3,5 +3,5 @@ package com.example.cupcycle.repository;
 import com.example.cupcycle.entity.Cafe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CafeRepository extends JpaRepository<Cafe, Long> {
+public interface CafeRepository extends JpaRepository<Cafe, Integer> {
 }
